dataset refactored into three modules
