+++ neuralmonkey/dataset/__init__.py
@@ -1,0 +1,3 @@
+from .helpers import from_files, load_dataset_from_files
+from .dataset import Dataset
+from .lazy_dataset import LazyDataset
